Se finaliza entrega

+++ Entregable-11-Orciuolo/src/controllers/products.controller.js
@@ -12,19 +12,27 @@ export default class ProductController {
    static async getByID(pid) {
       const product = await productsService.getByID(pid);
       if (!product) {
-         console.log('No es posible modificar el ID delproducto.');
+         console.log('No se encuentra el producto.');
          return (400);
       }
       return product;
    }
 
    static async updateByID(pid, data) {
-      await productsService.getByID(pid);
+      const product = await productsService.getByID(pid);
+      if (!product) {
+         console.log('No es posible modificar el producto.');
+         return (400);
+      }
       return productsService.updateByID(pid, data)
    }
 
    static async deleteByID(pid) {
-      await productsService.getByID(pid);
+      const product = await productsService.getByID(pid);
+      if (!product) {
+         console.log('No es posible eliminar el producto.');
+         return (400);
+      }
       return productsService.deleteByID(pid)
    }
 }
